[tools] refactor IO into subpackage
